Use lib in quantum wire
