--- year_4/TUM_quantum_sim/QW_sim/system.py
@@ -1,15 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class System:
-
-    Lx: float
-    Ly: float
-
-    wire_x: float
-    wire_y: float
-
-    Nx: int
-    Ny: int
-
-    x: float
